requirements.txt
fastapi
uvicorn
pydantic
httpx
python-dotenv
aiofiles
python-multipart
presidio-analyzer
presidio-anonymizer
opencv-python-headless
pillow
pytesseract
scikit-learn
pandas
joblib
groq
openai
numpy
requests